Implement signal support

+++ cmd/runhcs/kill.go
@@ -1,7 +1,14 @@
 package main
 
 import (
+	"strconv"
+	"strings"
+
 	"github.com/Microsoft/hcsshim/internal/appargs"
+	"github.com/Microsoft/hcsshim/internal/guestrequest"
+	"github.com/Microsoft/hcsshim/internal/hcs"
+	"github.com/Microsoft/hcsshim/internal/schema1"
+	"github.com/pkg/errors"
 	"github.com/urfave/cli"
 )
 
@@ -26,6 +33,7 @@ signal to the init process of the "ubuntu01" container:
 		if err != nil {
 			return err
 		}
+		defer c.Close()
 		status, err := c.Status()
 		if err != nil {
 			return err
@@ -34,9 +42,22 @@ signal to the init process of the "ubuntu01" container:
 			return errContainerStopped
 		}
 
-		sigstr := context.Args().Get(1)
-		if sigstr == "" {
-			sigstr = "SIGTERM"
+		signalsSupported := false
+		if c.IsHost {
+			uvm, err := hcs.OpenComputeSystem(vmID(c.ID))
+			if err != nil {
+				return err
+			}
+			defer uvm.Close()
+			if props, err := uvm.Properties(schema1.PropertyTypeGuestConnection); err == nil &&
+				props.GuestConnectionInfo.GuestDefinedCapabilities.SignalProcessSupported {
+				signalsSupported = true
+			}
+		}
+
+		signal, err := validateSigstr(context.Args().Get(1), signalsSupported, c.Spec.Linux != nil)
+		if err != nil {
+			return err
 		}
 
 		var pid int
@@ -49,6 +70,78 @@ signal to the init process of the "ubuntu01" container:
 			return err
 		}
 		defer p.Close()
-		return p.Kill() // BUGBUG: should be Signal
+
+		if signalsSupported {
+			opts := guestrequest.SignalProcessOptions{
+				Signal: signal,
+			}
+			return p.Signal(opts)
+		}
+
+		// Legacy signal issue a kill
+		return p.Kill()
 	},
+}
+
+func validateSigstr(sigstr string, signalsSupported bool, isLcow bool) (int, error) {
+	errInvalidSignal := errors.Errorf("invalid signal '%s'", sigstr)
+
+	// All flavors including legacy default to SIGTERM on LCOW CtrlC on Windows
+	if sigstr == "" {
+		if isLcow {
+			return 0xf, nil
+		}
+		return 0, nil
+	}
+
+	sigstr = strings.ToUpper(sigstr)
+
+	if !signalsSupported {
+		if isLcow {
+			switch sigstr {
+			case "15":
+				fallthrough
+			case "TERM":
+				fallthrough
+			case "SIGTERM":
+				return 0xf, nil
+			default:
+				return 0, errInvalidSignal
+			}
+		}
+		switch sigstr {
+		case "0":
+			fallthrough
+		case "CTRLC":
+			return 0x0, nil
+		default:
+			return 0, errInvalidSignal
+		}
+	}
+
+	var sigmap map[string]int
+	if isLcow {
+		sigmap = signalMapLcow
+	} else {
+		sigmap = signalMapWindows
+	}
+
+	signal, err := strconv.Atoi(sigstr)
+	if err != nil {
+		// Signal might still match the string value
+		for k, v := range sigmap {
+			if k == sigstr {
+				return v, nil
+			}
+		}
+		return 0, errInvalidSignal
+	}
+
+	// Match signal by value
+	for _, v := range sigmap {
+		if signal == v {
+			return signal, nil
+		}
+	}
+	return 0, errInvalidSignal
 }
